Port vendored libraries to Python 3 (compile-clean)
Migrate the vendored openpgp, bittorrent, urlgrabber, sets, pypack and
pylauncher trees to Python 3 syntax/semantics so every module compiles
cleanly under Python 3 (print/except/raise, long and octal literals,
has_key, xrange, iteritems, cmp/sort, stdlib renames such as
StringIO->io, urllib2->urllib.request, types module, etc.).

Note: these libraries now compile but are not yet fully runtime-correct
under Python 3. Known remaining work (tracked separately): implicit
relative imports in openpgp/bittorrent, bytes/str handling in the
bittorrent bencode/wire protocol and hashing, and a Python 3 rewrite of
modulegraph bytecode scanner used by the pypack freezer.

+++ src/bittorrent/Connecter.py
@@ -6,7 +6,7 @@ from binascii import b2a_hex
 from CurrentRateMeasure import Measure
 
 def toint(s):
-    return long(b2a_hex(s), 16)
+    return int(b2a_hex(s), 16)
 
 def tobinary(i):
     return (chr(i >> 24) + chr((i >> 16) & 0xFF) +
